Added TypeScript: The Complete Developer's Guide to Paid Courses
In my personal opinion, this is the one of the best TypeScript course that is out there.

Stephen Grider, the instructor, really goes through the ins and outs of TypeScript in a very engaging way. He covered popular design patterns and building complex projects using them. React and Express integration with TypeScript is also included.